🏷️rename: TemplateItem => InvitationItem
